'C++'-implementation of rectangularRange_HSV()

'C++'-implementation of rectangularRange_HSV()

Usage

rectangularRange_HSV_cpp(
  H,
  S,
  V,
  upper_bound,
  lower_bound,
  image_width,
  check_V
)

Arguments

H

respective component of a pixel.array

S

respective component of a pixel.array

V

respective component of a pixel.array

upper_bound

EITHER:

list of upper HSV-bounds, e.g. list(green = c(H_green_lower,S_green_lower,V_green_lower),drought = c(H_drought_lower,S_drought_lower,V_drought_lower))

single vector of length 3 declaring a set of HSV-values

lower_bound

see upper_bound

image_width

Width of pixel.array, as returned via dim(pixel.array)[1]

check_V

boolean toggle to also check the VALUE-component of an HSV-pixel

Value

A list-object with the following elements (when supplying one one pair of bounds)

pixel.idx - pixel-locations of pixels detected between lower and upper bound.

Upon failure to find any matching pixels, an empty matrix of dimensions [0, 1:2] is returned.

Note

The use of rectangularRange_HSV() is strongly discouraged in favour of this function, due to its drastically slower execution.

See also